improve first_element_in_dn

+++ src/entry.h
@@ -3,6 +3,7 @@
 #define ENTRY_H
 
 #include <stdbool.h>
+#include <stddef.h>
 
 // for use with stb array functions
 typedef char** STR_ARRAY;
@@ -32,7 +33,7 @@ STR_ARRAY entry_get_attribute(entry* e, const char* key);
 void entry_delete(entry* e);
 bool entry_attribute_exists(entry* e, const char* key, const char* value);
 bool entry_is_container(entry* e);
-void first_element_in_dn(char* buffer, const char* dn, int buffer_size);
+void first_element_in_dn(char* buffer, const char* dn, size_t buffer_size);
 bool entry_edit_value(entry* e, char* key, char* new_value);
 
 #endif
